% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.r
\name{assign_sample_groups}
\alias{assign_sample_groups}
\title{Assign sample groups based on terms}
\usage{
assign_sample_groups(values, control_terms, case_terms)
}
\arguments{
\item{values}{Vector of values to classify}

\item{control_terms}{Terms indicating controls}

\item{case_terms}{Terms indicating cases}
}
\value{
Vector of group assignments
}
\description{
Assign sample groups based on terms
}
\keyword{internal}
