% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-pseudo.R
\name{avg.transform}
\alias{avg.transform}
\title{Averaging transformation}
\usage{
avg.transform(x, log.transform = TRUE)
}
\arguments{
\item{x}{vector of feature counts}

\item{log.transform}{whether or not log transform should be applied}
}
\description{
Averaging transformation for a vector of feature counts.
For more details, see "Good-Turing Smoothing Without Tears", by William A. Gale.
}

