% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedEPICv2.R
\docType{data}
\name{bedEPICv2}
\alias{bedEPICv2}
\title{BED EPICv2 probes}
\format{
BED file.
}
\usage{
data(bedEPICv2)
}
\description{
This dataset contains the EPICv2 probes coordinates.
}
\examples{
\dontrun{
data(bedEPICv2)
head(bedEPICv2)
}
}
\keyword{datasets}
