% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.r
\name{create_geo_sample_sheet_with_mappings}
\alias{create_geo_sample_sheet_with_mappings}
\title{Create sample sheet with user-defined group mappings}
\usage{
create_geo_sample_sheet_with_mappings(
  pheno_data,
  idat_files,
  group_column,
  group_mappings,
  max_samples
)
}
\arguments{
\item{pheno_data}{Phenotypic data from GEO}

\item{idat_files}{Vector of IDAT file paths}

\item{group_column}{Column name for groups}

\item{group_mappings}{Named list mapping values to groups}

\item{max_samples}{Maximum samples to include}
}
\value{
Data frame with sample sheet
}
\description{
Create sample sheet with user-defined group mappings
}
\keyword{internal}
