% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.r
\name{create_geo_zip}
\alias{create_geo_zip}
\title{Create ZIP file for GIMP processing}
\usage{
create_geo_zip(idat_dir, base_dir, geo_id)
}
\arguments{
\item{idat_dir}{Directory containing organized IDAT files}

\item{base_dir}{Base directory for ZIP creation}

\item{geo_id}{GEO ID for naming}
}
\value{
Path to created ZIP file
}
\description{
Create ZIP file for GIMP processing
}
\keyword{internal}
