% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ICR_matrices.R
\name{create_ICR_matrices}
\alias{create_ICR_matrices}
\title{Create the Imprinted Matrix}
\usage{
create_ICR_matrices(myCombat, bedEPIC_file = "bedEPIC.csv")
}
\description{
Create the Imprinted Matrix
}
\examples{
plot_heat(score_data, "plots/heat_percentiles.png")
}
