% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.r
\name{detect_array_platform}
\alias{detect_array_platform}
\title{Detect array platform from GEO metadata}
\usage{
detect_array_platform(platforms, title = "", summary = "")
}
\arguments{
\item{platforms}{Vector of platform IDs}

\item{title}{Dataset title}

\item{summary}{Dataset summary}
}
\value{
Detected array type
}
\description{
Detect array platform from GEO metadata
}
\keyword{internal}
