% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\docType{data}
\name{dm.fit.lung}
\alias{dm.fit.lung}
\title{Lung microbiome data}
\format{A list with five objects}
\description{
A list with 5 objects, including \code{gamma}, the parameter vector for a
Dirichlet distribution. Estimated from the (non-smoker) lung data set described in
Charlson ES et al.: Topographical continuity of bacterial populations in the healthy
human respiratory tract. Am J Respir Crit Care Med 2011, 184:957-963.
Estimated using package \code{dirmult}.
}
\details{
\itemize{
   \item gamma. Parameter vector for Dirichlet distribution
   \item pi. Mean vector for Dirichlet distribution
   \item loglik. Final log-likelihood value
   \item ite. Number of iterations used
   \item theta. Estimated theta-value
}
}

