% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr19.R
\docType{data}
\name{dmr19}
\alias{dmr19}
\title{Imprinted Regions}
\format{
A data frame with iDMRs coordinates.
}
\usage{
data(dmr19)
}
\description{
This dataset contains the Human Imprinted regions coordinates in hg19.
}
\examples{
data(dmr19)
head(dmr19)
}
\keyword{datasets}
