% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmr38.R
\docType{data}
\name{dmr38}
\alias{dmr38}
\title{Imprinted Regions}
\format{
A data frame with iDMRs coordinates.
}
\usage{
data(dmr38)
}
\description{
This dataset contains the Human Imprinted regions coordinates in hg38.
}
\examples{
data(dmr38)
head(dmr38)
}
\keyword{datasets}
