% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRs.hg38.R
\docType{data}
\name{DMRs.hg38}
\alias{DMRs.hg38}
\title{Imprinted Regions}
\format{
A data frame with iDMRs coordinates.
}
\usage{
data(DMRs.hg38)
}
\description{
This dataset contains the Human Imprinted regions coordinates in hg38.
}
\examples{
\dontrun{
data(DMRs.hg38)
head(DMRs.hg38)
}
}
\keyword{datasets}
