% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb-pseudo.R
\name{eb.pseudo}
\alias{eb.pseudo}
\title{empirical Bayes adjustment of sparse counts matrix}
\usage{
eb.pseudo(counts)
}
\arguments{
\item{counts}{Integer count matrix with features in rows and samples in columns}
}
\description{
\code{eb.pseudo} performs empirical Bayes adjustment for sparse count matrix "counts",
where each sample is independently adjusted.
}
\details{
Essentially a wrapper funtion for \code{eb.pseudo.sample}, which performs empirical
Bayes adjustment for a feature counts vector for a single sample.
}
\seealso{
\code{\link{eb.pseudo.sample}} for adjustment of a single sample. See
   \code{\link{sim.groups}} for a simple example.
}

