% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb-pseudo.R
\name{eb.pseudo.sample}
\alias{eb.pseudo.sample}
\title{empirical Bayes adjustment of sparse counts vector}
\usage{
eb.pseudo.sample(x)
}
\arguments{
\item{x}{Feature count vector}
}
\value{
List with objects \code{z}, \code{psi}, and \code{mu}
}
\description{
Primarily a helper function for \code{eb.pseudo}, but can be used
   on its own to normalize a single sample
}
\seealso{
\code{\link{eb.pseudo}}
}

