% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.r
\name{extract_compressed_files}
\alias{extract_compressed_files}
\title{Extract compressed files in a directory}
\usage{
extract_compressed_files(dir_path)
}
\arguments{
\item{dir_path}{Directory to search for compressed files}
}
\value{
NULL (invisibly). Function is called for its side effects of extracting files.
}
\description{
Extract compressed files in a directory
}
\keyword{internal}
