% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.r
\name{extract_sample_id_from_idat}
\alias{extract_sample_id_from_idat}
\title{Extract sample ID from IDAT filename}
\usage{
extract_sample_id_from_idat(idat_basename)
}
\arguments{
\item{idat_basename}{IDAT file basename}
}
\value{
Potential sample ID
}
\description{
Extract sample ID from IDAT filename
}
\keyword{internal}
