% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.r
\name{extract_sentrix_info}
\alias{extract_sentrix_info}
\title{Extract Sentrix information from IDAT basename}
\usage{
extract_sentrix_info(idat_basenames)
}
\arguments{
\item{idat_basenames}{Vector of IDAT basenames}
}
\value{
List with Sentrix ID and position
}
\description{
Extract Sentrix information from IDAT basename
}
\keyword{internal}
