% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb-pseudo.R
\name{fac.moment}
\alias{fac.moment}
\title{calculates the first two factorial moments of a numeric vector}
\usage{
fac.moment(x, return.scaled = TRUE)
}
\arguments{
\item{x}{Integer count matrix with features in rows and samples in columns}

\item{return.scaled}{Whether or not to return the scaled factorial moments}
}
\description{
Helper function for \code{eb.pseudo.sample}
}
\seealso{
\code{\link{eb.pseudo.sample}}
}

