% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.r
\name{find_and_organize_idats}
\alias{find_and_organize_idats}
\title{Find and organize IDAT files}
\usage{
find_and_organize_idats(source_dir, target_dir)
}
\arguments{
\item{source_dir}{Directory to search for IDAT files}

\item{target_dir}{Directory to copy organized IDAT files}
}
\value{
Vector of IDAT file paths
}
\description{
Find and organize IDAT files
}
\keyword{internal}
