% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-pseudo.R
\name{gt.fit.lm}
\alias{gt.fit.lm}
\title{Linear model for average transformed counts}
\usage{
gt.fit.lm(x)
}
\arguments{
\item{x}{vector of feature counts}
}
\description{
Low-level function for smoothed Good-Turing frequency estimation. Fits
   a linear model to counts subjected to "averaging transformation" in order to obtain
   better frequency estimates for samples where no 1s are observed.
}
\seealso{
\code{\link{avg.transform}}
}

