% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-pseudo.R
\name{gt.pseudo}
\alias{gt.pseudo}
\title{Good-Turing adjustment of sparse counts matrix}
\usage{
gt.pseudo(counts)
}
\arguments{
\item{counts}{Integer count matrix with features in rows and samples in columns}
}
\description{
\code{gt.pseudo} performs Good-Turing adjustment for sparse count matrix "counts",
where each sample is independently adjusted.
}
\details{
Essentially a wrapper funtion for \code{gt.pseudo.sample}, which performs
Good-Turing adjustment for a feature counts vector for a single sample.
}
\seealso{
\code{\link{gt.pseudo.sample}} for adjustment of a single sample. See
   \code{\link{sim.groups}} for a simple example.
}

