% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-pseudo.R
\name{gt.pseudo.sample}
\alias{gt.pseudo.sample}
\title{Good-Turing adjustment of sparse counts vector}
\usage{
gt.pseudo.sample(x)
}
\arguments{
\item{x}{Feature count vector}
}
\description{
Primarily a helper function for \code{gt.pseudo}, but can be used
   on its own to normalize a single sample
}
\seealso{
\code{\link{gt.pseudo}}
}

