% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_functions.r
\name{match_geo_samples}
\alias{match_geo_samples}
\title{Match samples between phenotypic data and IDAT files}
\usage{
match_geo_samples(pheno_data, idat_basenames)
}
\arguments{
\item{pheno_data}{Phenotypic data}

\item{idat_basenames}{IDAT file basenames}
}
\value{
Data frame with matched samples
}
\description{
Match samples between phenotypic data and IDAT files
}
\keyword{internal}
