% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CpG_coverage.R
\name{plot_CpG_coverage}
\alias{plot_CpG_coverage}
\title{Plot ICR CPG Matrix}
\usage{
plot_CpG_coverage(df_ICR_cpg, bedmeth = "v1")
}
\arguments{
\item{df_ICR_cpg}{A data frame containing CpG counts for ICR regions. Each row represents a different CpG probe, and columns contain sample-related information.}

\item{bedmeth}{A character string specifying the BED data version to use for mapping CpG coverage. Options are `"v1"` (EPIC v1), `"v2"` (EPIC v2), or `"450k"` (450k array). Default is `"v1"`.}
}
\value{
A plot visualizing the CpG coverage for ICR regions based on the specified BED data.
}
\description{
This function plots the CpG coverage for Imprinted Control Regions (ICRs) using the provided data frame of CpG counts. It compares CpG counts in the specified BED data version for visual analysis.
}
\examples{
plot_CpG_coverage(df_ICR_cpg_counts, bedmeth = "v1")
}
