% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line_ICR.R
\name{plot_line_ICR}
\alias{plot_line_ICR}
\title{Plot Line Plot for ICR Methylation}
\usage{
plot_line_ICR(significantDMPs, ICRcpg, ICR, sampleInfo, interactive = TRUE)
}
\arguments{
\item{significantDMPs}{A data frame containing information about significant DMPs. Must include columns `ICR`, `start`, and `end`.}

\item{ICRcpg}{A data frame or matrix containing CpG methylation data. Includes CpG coordinates (`cstart`) and methylation values.}

\item{ICR}{A character string specifying the name of the ICR region to be plotted.}

\item{sampleInfo}{A character vector providing group labels (e.g., `"Control"` or `"Case"`) for each sample in the methylation data.}

\item{interactive}{A logical value indimessageing whether to return an interactive `plotly` plot (`TRUE`) or a static `ggplot2` plot (`FALSE`). Default is `TRUE`.}
}
\value{
A plot representing the line plot of methylation values across the specified ICR region, highlighting significant DMPs. 
The plot is either a `ggplot2` object or a `plotly` object, depending on the value of `interactive`.
}
\description{
This function generates a line plot to visualize methylation values across a specified ICR.
Users can choose between a static `ggplot2` plot or an interactive `plotly` plot.
}
\examples{
\dontrun{
# Example data for significantDMPs
plot <- plot_line_ICR(significantDMPs, ICRcpg, ICR = "KCNQ1OT1:TSS-DMR", sampleInfo = sampleInfo, interactive = T)
print(plot)
}
}
