% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line_region.R
\name{plot_line_region}
\alias{plot_line_region}
\title{Plot Line Plot for Imprinted DMR Methylations}
\usage{
plot_line_region(significantDMPs, ICRcpg, ICR, sampleInfo, interactive = TRUE)
}
\arguments{
\item{significantDMPs}{A data frame containing information about significant DMPs. Must include columns `ICR`, `start`, and `end`.}

\item{ICRcpg}{A data frame or matrix containing CpG methylation data. Includes CpG coordinates (`cstart`) and methylation values.}

\item{ICR}{A character string specifying the name of the ICR region to be plotted.}

\item{sampleInfo}{A character vector providing group labels (e.g., `"Control"` or `"Case"`) for each sample in the methylation data.}

\item{interactive}{A logical value indicating whether to return an interactive `plotly` plot (`TRUE`) or a static `ggplot2` plot (`FALSE`). Default is `TRUE`.}
}
\value{
A plot representing the line plot of methylation values across the specified ICR region, highlighting significant DMPs. 
The plot is either a `ggplot2` object or a `plotly` object, depending on the value of `interactive`.
}
\description{
This function generates a line plot to visualize methylation values across a specified Imprinted Differentially Methylated Region (ICR).
Users can choose between a static `ggplot2` plot or an interactive `plotly` plot.
}
\examples{
# Example data for significantDMPs
significantDMPs <- data.frame(
  ICR = c("ICR1", "ICR1", "ICR2"),
  start = c(100, 200, 150),
  end = c(150, 250, 200)
)

# Example data for ICRcpg
ICRcpg <- data.frame(
  CpG1 = runif(10, 0, 1), CpG2 = runif(10, 0, 1),
  cstart = seq(100, 190, by = 10)
)

# Example sample info
sampleInfo <- c("Control", "Case")

# Plot methylation for a specific ICR region
plot <- plot_line_region(significantDMPs, ICRcpg, ICR = "ICR1", sampleInfo = sampleInfo, interactive = TRUE)
plot
}
