% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-pseudo.R
\name{pred.mass.smooth}
\alias{pred.mass.smooth}
\title{Simple Good-Turing frequency estimator, smoothed}
\usage{
pred.mass.smooth(x, k = 0)
}
\arguments{
\item{x}{vector of feature counts}

\item{k}{count value for which mass should be estimated}
}
\description{
Estimates the mass which should be assigned to all features with count
   k in sample x, smoothed by Gale's method (see "Good-Turing Smoothing Without Tears")
}

