% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.counts}
\alias{sim.counts}
\title{Dirichlet-multinomial counts simulation}
\usage{
sim.counts(dm.fit, ls, remove.zero.rows = FALSE)
}
\arguments{
\item{dm.fit}{\code{dirmult} object. Alternatively, a list object with at least \code{gamma} object,
corresponding to parameters of a Dirichlet distribution}

\item{ls}{vector of library sizes of samples to be generated}

\item{remove.zero.rows}{whether or not to remove rows of all zeroes}
}
\description{
Simulates counts according to a Dirichlet-multinomial model, according
to model parameters given by object dm.fit.
}
\details{
Simulates a single set of counts, generated according to the
parameters in object \code{dm.fit}. Requires package \code{dirmult} to generate samples.
}

