% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.groups}
\alias{sim.groups}
\title{Dirichlet-multinomial case/control counts simulation}
\usage{
sim.groups(dm.fit, Nshuffle = 15, params = list(mu1 = 300, sd1 = 50, N1 =
  20, mu2 = 300, sd2 = 50, N2 = 20))
}
\arguments{
\item{dm.fit}{\code{dirmult} object. Alternatively, a list object with at least \code{gamma} object,
corresponding to parameters of a Dirichlet distribution}

\item{Nshuffle}{number of features to shuffle}

\item{params}{list of parameters describing the number the mean and standard deviation of
library sizes for case and control samples, as well as the number of samples of each}
}
\value{
List with objects \code{counts}, \code{i.spike}, \code{groups}: a counts table, feature
   names which have been shuffled, and group membership (case/control), respectively
}
\description{
Simulates counts according to a Dirichlet-multinomial model, according
to model parameters given by object dm.fit.
}
\details{
Simulates two sets of counts (case/control), where one set is generated
according to the parameters in object \code{dm.fit}, and and the second set is
generated such that \code{Nshuffle} number of high abundance features are
shuffled with \code{Nshuffle} number of low abundance features, generating a final
data set with a large number of differences between case and control samples for
large enough \code{Nshuffle}. Requires package \code{dirmult} to generate samples.
}
\examples{
   library(HMP)
   library(dirmult)
   S <- sim.groups(dm.fit.eso)
   S$eb <- eb.pseudo(S$counts)
   S$gt <- gt.pseudo(S$counts)
   round(head(S$eb),4)

   S <- sim.groups(dm.fit.lung, Nshuffle = 50,
        params = list(mu1 = 2000, sd1 = 200, N1 = 20,
                      mu2 = 6000, sd2 = 600, N2 = 20))
   S$eb <- eb.pseudo(S$counts)
   S$gt <- gt.pseudo(S$counts)
   round(head(S$gt),4)
}
\seealso{
\code{\link{sim.counts}} for simulation of a single group of samples
}

