% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed450k.R
\docType{data}
\name{bed450k}
\alias{bed450k}
\title{BED 450K probes}
\format{
BED file.
}
\usage{
data(bed450k)
}
\description{
This dataset contains the 450K array probes coordinates.
}
\examples{
\dontrun{
data(bed450k)
head(bed450k)
}
}
\keyword{datasets}
