% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedEPICv1.R
\docType{data}
\name{bedEPICv1}
\alias{bedEPICv1}
\title{BED EPICv1 probes}
\format{
BED file.
}
\usage{
data(bedEPICv1)
}
\description{
This dataset contains the EPICv1 probes coordinates.
}
\examples{
data(bedEPICv1)
head(bedEPICv1)
}
\keyword{datasets}
