% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minfi_helpers.R
\name{check_minfi_functions}
\alias{check_minfi_functions}
\title{Check Available minfi Functions}
\usage{
check_minfi_functions()
}
\value{
Character vector of all available minfi functions
}
\description{
Diagnostic function to check which minfi functions are available in the current installation.
}
