% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMRs.hg19.R
\docType{data}
\name{DMRs.hg19}
\alias{DMRs.hg19}
\title{Imprinted Regions}
\format{
A data frame with iDMRs coordinates.
}
\usage{
data(DMRs.hg19)
}
\description{
This dataset contains the Human Imprinted regions coordinates in hg19.
}
\examples{
data(DMRs.hg19)
head(DMRs.hg19)
}
\keyword{datasets}
