% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMR_heatmap.R
\name{iDMR_heatmap}
\alias{iDMR_heatmap}
\title{Generate Heatmap of Imprinted DMRs Methylation}
\usage{
iDMR_heatmap(
  df_ICR,
  group_vector,
  control_label = "Control",
  case_label = "Case",
  bedmeth = "v1",
  cluster_by = "cord",
  annotation_col = NULL
)
}
\arguments{
\item{df_ICR}{A data frame or matrix containing methylation beta values for Imprinted DMRs.}

\item{group_vector}{A vector indicating the group labels (e.g., "Control" and "Case") for each sample in `df_ICR`.
Each element in `group_vector` should correspond to a sample in `df_ICR`.}

\item{control_label}{A character string specifying the label for the control group in `group_vector`. Default is `"Control"`.}

\item{case_label}{A character string specifying the label for the case group in `group_vector`. Default is `"Case"`.}

\item{bedmeth}{A character string specifying the BED data version for DMR coordinates. Options are `"v1"`, `"v2"`, or `"450k"`. Default is `"v1"`.}

\item{cluster_by}{A character string specifying the clustering method for rows in the heatmap. Options are `"cord"` or `"meth"`. Default is `"cord"`.}

\item{annotation_col}{A named list of colors for each unique value in `group_vector`. If `NULL`, default colors are assigned using the "viridis" palette. Default is `NULL`.}
}
\value{
A heatmap plot visualizing methylation of Imprinted DMRs.
}
\description{
This function generates a heatmap for visualizing methylation data of Imprinted Differentially Methylated Regions (DMRs).
It allows custom color schemes for group annotations, with default colors from the "viridis" palette.
}
\examples{
# Example group_vector with "Case" and "Control" labels for each sample
group_vector <- c(rep("Case", 10), rep("Control", 10))
iDMR_heatmap(df_ICR = my_ICR_data, group_vector = group_vector, annotation_col = list(Sample = c("darkgreen", "darkred")))
}
