% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ICRs.R
\name{make_ICRs}
\alias{make_ICRs}
\title{Create the ICR Matrix}
\usage{
make_ICRs(Bmatrix, bedmeth = "v1")
}
\arguments{
\item{Bmatrix}{A data frame or matrix containing methylation beta values. Rows should represent CpG probes, and columns represent samples.}

\item{bedmeth}{A character string indicating the BED data version to use for CpG mapping. Options are `"v1"` (EPIC v1), `"v2"` (EPIC v2), or `"450k"` (450k array). Default is `"v1"`.}
}
\value{
A data frame representing the ICR matrix, structured by CpG probes and samples.
}
\description{
This function generates an ICR (Imprinted Control Region) matrix from a given beta matrix, using specified BED data for CpG mapping. The ICR matrix provides data organized by CpG probes and samples. The coordinates of the Human Imprinted regions are taken from https://doi.org/10.1080/15592294.2016.1264561
}
\examples{
ICRmatrix <- make_ICRs(Bmatrix = df, bedmeth = "v1")
}
