% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cpgs_coverage.R
\name{plot_cpgs_coverage}
\alias{plot_cpgs_coverage}
\title{Plot ICR CpG Matrix with Counts and Percentage Coverage}
\usage{
plot_cpgs_coverage(df_ICR_cpg, bedmeth = "v1")
}
\arguments{
\item{df_ICR_cpg}{A data frame containing CpG counts for ICR regions. Each row represents a different CpG probe, and columns contain sample-related information.}

\item{bedmeth}{A character string specifying the BED data version to use for mapping CpG coverage. Options are `"v1"` (EPIC v1), `"v2"` (EPIC v2), or `"450k"` (450k array). Default is `"v1"`.}
}
\value{
A list containing two plots (counts and percentage coverage) and the data frame with CpG counts and coverage information.
}
\description{
This function plots the CpG coverage for Imprinted Control Regions (ICRs) using the provided data frame of CpG counts.
It compares CpG counts in the specified BED data version for visual analysis and includes an additional plot for percentage coverage.
}
\examples{
plot_cpgs_coverage(df_ICR_cpg_counts, bedmeth = "v1")
}
